requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
motor==3.7.0
passlib[argon2]==1.7.4
slowapi==0.1.9
python-multipart==0.0.20
python-jose==3.3.0
redis>=4.5
pydantic==2.10.6
pydantic[email]==2.10.6
python-dotenv==1.0.1
httpx==0.28.1

# Testing dependencies
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
